+++ ConsoleApplication1/Individual.cpp
@@ -1,5 +1,9 @@
 #include "pch.h"
 #include <random>
+#include <cstdlib>
+#include <cstdlib>
+#include <cstdlib>
+#include <cstdlib>
 
 static const unsigned genomeShapeData[] = { 28 * 28, 7 * 7, 10, 1 };
 const unsigned* Individual::genomeShape = genomeShapeData;
@@ -9,7 +13,7 @@ int Individual::lastId = -1;
 
 void Individual::createADN()
 {
-	for (size_t i = 0; i < genomeSize; i++)
+	for (unsigned i = 0; i < genomeSize; i++)
 	{
 		ADN[i] = (rand() / (double)RAND_MAX) * 2 - 1;
 	}
@@ -23,12 +27,12 @@ void Individual::createGenome()
 	genome = new Matrix[genomeShapeLength];
 
 	int index = 0;
-	for (size_t i = 0; i < genomeShapeLength - 1; i++)
+	for (unsigned i = 0; i < genomeShapeLength - 1; i++)
 	{
 		genome[i] = Matrix(genomeShape[i + 1], genomeShape[i]);
-		for (size_t j = 0; j < genomeShape[i + 1]; j++)
+		for (unsigned j = 0; j < genomeShape[i + 1]; j++)
 		{
-			for (size_t k = 0; k < genomeShape[i]; k++)
+			for (unsigned k = 0; k < genomeShape[i]; k++)
 			{
 				genome[i].set(j, k, ADN[index++]);
 			}
@@ -41,7 +45,7 @@ void Individual::mutate(const Individual& A, const Individual& B)
 	bool parentA = true;
 	int nbCuts = genomeSize / 100;
 
-	for (size_t i = 0; i < genomeSize; i++)
+	for (unsigned i = 0; i < genomeSize; i++)
 	{
 		if (rand() % nbCuts == 0)
 			parentA = !parentA;
@@ -70,7 +74,7 @@ Individual::Individual(const Individual & other)
 	id = other.id;
 
 	ADN = new double[genomeSize];
-	for (size_t i = 0; i < genomeSize; i++)
+	for (unsigned i = 0; i < genomeSize; i++)
 		ADN[i] = other.ADN[i];
 
 	createGenome();
@@ -89,7 +93,7 @@ Matrix Individual::getResults(const Matrix& inputs)
 {
 	Matrix secondMatrix = inputs;
 
-	for (size_t i = 0; i < genomeShapeLength - 1; i++)
+	for (unsigned i = 0; i < genomeShapeLength - 1; i++)
 	{
 		secondMatrix = genome[i] * secondMatrix;
 		secondMatrix.matrixTanh();
